feat: refactoring for code re-usage
(cherry picked from commit 44f45d41e70f65db542c6a9396a0cb2948dfb1f3)
