+++ src/runtime/components/ShopwareFrontendsPos.vue
@@ -1,9 +1,10 @@
 <script setup lang="ts">
-import { useMollie, useAsyncData, useShopwareContext, useUser } from '#imports'
+import { useMollie, useAsyncData, useShopwareContext, useUser, watch } from '#imports'
 import { onMounted, ref, type Ref } from 'vue'
 import type { MolliePosTerminal, MollieLocale, ShopwareLocale } from '../../types'
 import { ApiClientError } from '@shopware/api-client'
 import { shopwareLocaleToMollieLocale } from '../utils/localeTransformer'
+import { useShopwareMollie } from '../composables/useShopwareMollie'
 
 const emits = defineEmits<{
     (e: 'get-terminals-error', error: string | undefined): void
@@ -18,28 +19,12 @@ const props = defineProps<{
 }>()
 
 const { apiClient } = useShopwareContext()
-
-const { data: mollieConfig } = await useAsyncData('mollieConfig', async () => {
-    try {
-        const config = await apiClient.invoke('getConfig get /mollie/config')
-        // use the locale from the props if it exists, otherwise use the locale from the mollie config
-        // the locale-code from Shopware is in another format than the one from Mollie, so those have to be aligned.
-        if (config.data) {
-            const localeFromShopware = config.data.locale as ShopwareLocale
-            const mollieLocale = shopwareLocaleToMollieLocale(localeFromShopware)
-            config.data.locale = props.locale ?? mollieLocale
-        }
-        return config
-    } catch (error) {
-        if (error instanceof ApiClientError) {
-            console.error(error)
-        } else {
-            console.error('==>', error)
-        }
-    }
+const { mollieConfig } = useShopwareMollie({ locale: props.locale })
+const { init } = useMollie()
+watch(mollieConfig, () => {
+    init(mollieConfig.value)
 })
 
-const { init } = useMollie(mollieConfig.value)
 const { user } = useUser()
 
 const terminals: Ref<MolliePosTerminal[] | []> = ref([])
@@ -78,7 +63,6 @@ const savePosTerminal = async () => {
 }
 
 onMounted(async () => {
-    await init()
     await getPosTerminals()
 })
 </script>
